refactor: modularize plugin services, models, serializers, and typing
Extract backend HTTP/SSE sync logic into services/backend_proxy and
keepalive checks into services/service_status. Split models and API
serializers into focused modules with stable package exports. Add
type_defs and tighten utils; document helpers and update tests.
