+++ netbox_proxbox/services/__init__.py
@@ -1,0 +1,19 @@
+"""Service-layer modules (backend HTTP proxy, keepalive checks)."""
+
+from netbox_proxbox.services.backend_proxy import (
+    get_fastapi_request_context,
+    iter_backend_sse_lines,
+    sse_error_frames,
+    sync_full_update_resource,
+    sync_resource,
+)
+from netbox_proxbox.services.service_status import ServiceStatus
+
+__all__ = (
+    "ServiceStatus",
+    "get_fastapi_request_context",
+    "iter_backend_sse_lines",
+    "sse_error_frames",
+    "sync_full_update_resource",
+    "sync_resource",
+)
