+++ tests/test_form_and_helper_source_contracts.py
@@ -43,7 +43,7 @@ def test_runtime_code_validates_fastapi_url_helper_payload_shape():
     runtime_files = [
         "netbox_proxbox/views/__init__.py",
         "netbox_proxbox/views/cards.py",
-        "netbox_proxbox/views/keepalive_status.py",
+        "netbox_proxbox/services/service_status.py",
         "netbox_proxbox/views/sync.py",
         "netbox_proxbox/websocket_client.py",
     ]
